pyproject.toml
[tool.poetry]
name = "leetcode-cn2us"
version = "0.1.0"
description = "Migrate the submission of leetcode-cn to us"
authors = ["nzooherd <nzooherd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
